+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.0-beta8] - 2026-04-?
+
+### Changed
+
+- Enhance Radarr and Sonarr ID resolution for movie and series candidates during tagging (if not yet cached from full scan)
+
 ## [0.1.0-beta7] - 2026-04-16
 
 ### Changed